backend/requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==1.1.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==1.1.8
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==1.4.2
Django==2.2.20
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-webpush==0.3.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
hiredis==2.0.0
http-ece==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
msgpack-python==0.5.6
psycopg2==2.8.6
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pytz==2020.5
pywebpush==1.9.4
redis==2.10.6
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.3
zope.interface==5.3.0
